agrovision-azure-func/ml-ndvi/requirements.txt
azure-functions
requests
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu
torchvision==0.6.0+cpu
azureml-defaults
azure-cosmos
azure-storage-blob